+++ ChangeLog.md
@@ -3,6 +3,9 @@ Match DSL ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Improved performance of type lookups by adding special case handling
+  for primitives to the `util.match.TypeOf` class.
+  (@thekid) 
 * Added new class, `util.match.KeyOf`. It works only with strings and
   integers but has higher performance than the `ValueOf` class.
   (@thekid)